requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-rest-auth==0.9.5
django-oauth-toolkit==1.5.0
drf-yasg==1.20.0
psycopg2-binary==2.8.6
djangorestframework-simplejwt==4.7.1
django-filter==2.4.0
pytest==6.2.4
pytest-django==4.1.0
